LIQ-138: Added entry database and added from preset selection

+++ app/src/main/java/com/tsquaredapplications/liquid/home/resources/HomeResourceWrapper.kt
@@ -4,5 +4,5 @@ import com.tsquaredapplications.liquid.setup.LiquidUnit
 
 interface HomeResourceWrapper {
 
-    fun getGoalProgressText(progress: Int, goal: Int, unit: LiquidUnit): String
+    fun getGoalProgressText(progress: Double, goal: Int, unit: LiquidUnit): String
 }
